requirements.txt
biopython
pandas
streamlit
matplotlib
plotly